+++ fcrepo-export/src/main/java/org/fcrepo/exporter/Importer.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.export;
+package org.fcrepo.exporter;
 
 /**
  * @author awoods
