+++ fcrepo-export/src/test/java/org/fcrepo/exporter/ArgParserTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.export;
+package org.fcrepo.exporter;
 
 import java.io.File;
 import java.net.URI;
