+++ fcrepo-export/src/test/java/org/fcrepo/exporter/ImportExportDriverTest.java
@@ -15,42 +15,41 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.export;
+package org.fcrepo.exporter;
+
+import junit.framework.Test;
+import junit.framework.TestCase;
+import junit.framework.TestSuite;
 
 /**
- * @author ruebot
- * @since 2016-08-29
+ * Unit test for simple App.
  */
-public class ImportExportDriver {
+public class ImportExportDriverTest
+    extends TestCase
+{
+    /**
+     * Create the test case
+     *
+     * @param testName name of the test case
+     */
+    public ImportExportDriverTest(String testName )
+    {
+        super( testName );
+    }
 
     /**
-     * The main entry point
-     *
-     * @param args from the command line
+     * @return the suite of tests being tested
      */
-    public static void main(final String[] args) {
-        final ImportExportDriver driver = new ImportExportDriver();
-
-        try {
-            driver.run(args);
-
-        } catch (final Exception e) {
-            // swallow exception
-        }
+    public static Test suite()
+    {
+        return new TestSuite( ImportExportDriverTest.class );
     }
 
-    private void run(final String[] args) {
-        final ArgParser parser = new ArgParser();
-        final Config config = parser.parse(args);
-
-        if (config.isImport()) {
-            final Importer importer = new Importer(config);
-            importer.run();
-
-        } else {
-            final Exporter exporter = new Exporter(config);
-            exporter.run();
-        }
+    /**
+     * Rigourous Test :-)
+     */
+    public void testexport()
+    {
+        assertTrue( true );
     }
-
 }
